+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## Version 1.1.0
+## Version 2.1
+- Changed `ApiSuccessPaginated` to split the pagination data.
+- Documentation builder changed to create swagger and postman files separately.
+- Fixed swagger syntax 
 
+## Version 2.0
+- Added documentation builder
+
+## Version 1.1.0
 - Changed `data` to `payload`
 - Fixed controller trait to handle `ValidationException` for json.